LayoutTests/editing/selection/5497643-expected.txt

ALERT: SUCCESS: The selection was cleared.
This tests to make sure that a selection inside a textarea is cleared when the textarea is removed from the document. Not clearing it led to crashes.
